Fix VS 32-bit warning: make downcast from uint64_t to size_t explicit

+++ src/lib_media/demux/gpac_demux_mp4_full.cpp
@@ -158,7 +158,7 @@ bool GPACDemuxMP4Full::safeProcessSample() {
 }
 
 void GPACDemuxMP4Full::process(Data data) {
-	reader->pushData(data->data(), data->size());
+	reader->pushData(data->data(), (size_t)data->size());
 
 	if (!reader->movie) {
 		if (!openData()) {
